fix(vt-daemon): project task folders in live state
