+++ packages/libraries/graph-state/package.json
@@ -13,6 +13,7 @@
     "./fixtures": "./src/fixtures.ts",
     "./fixtures/serialization": "./src/fixtures/serialization.ts",
     "./project": "./src/project.ts",
+    "./projectGraphDerivedFolderTree": "./src/projectGraphDerivedFolderTree.ts",
     "./state/folderVisibility/implicitRoots": "./src/state/folderVisibility/implicitRoots.ts",
     "./state/folderVisibilityStore": "./src/state/folderVisibilityStore.ts",
     "./state/selectionStore": "./src/state/selectionStore.ts",
